HWDEV-2079 move is_emergency fron can_controller to board_controller
