+++ lexxpluss_apps/src/can_controller.hpp
@@ -70,7 +70,6 @@ void init();
 void run(void *p1, void *p2, void *p3);
 bool get_emergency_switch();
 bool get_bumper_switch();
-bool is_emergency();
 extern k_thread thread;
 extern k_msgq /*msgq_bmu,*/ msgq_board, msgq_control;
